fastapi-base-ds/pyproject.toml
[tool.poetry]
name = "fastapi-base-ds"
version = "0.1.0"
description = "Proyecto base con FastAPI"
authors = ["Bruno <pazosbruno@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.112.0"}
sqlalchemy = "^2.0.31"
python-dotenv = "^1.0.1"
pytest = "^8.3.2"
httpx = "^0.27.0"
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
